requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==4.0.0
dj-static==0.0.6
Django==4.2.0
django-allauth==0.61.1
django-appconf==1.0.6
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.2.0
gunicorn==21.2.0
idna==3.6
kombu==5.3.5
oauthlib==3.2.2
packaging==24.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
